WIP: Prevent length mismatch of trigger and value
The README states the following:

    Does not trigger if the trigger is the same length as the value, so
    that abbreviations that correct typos do not remind of incorrect
    spelling.

The values in `abbr_data.value` seem to be preceded by whitespace. This
whitespace causes the length of these values to be one character longer
than that of `abbr_data.trigger`.

 ## For review

I have not investigated why this is the case, nor whether it would be
more appropriate to incorporate this change somehow into the
`abbremand.nvim` plugin instead.